refactor: 키워드로 검색한 투표 개수를 조회하도록 변경

+++ bucketback-api/src/main/java/com/programmers/bucketback/domains/vote/api/dto/response/VoteGetByKeywordResponse.java
@@ -8,7 +8,7 @@ import com.programmers.bucketback.domains.vote.model.VoteSummary;
 public record VoteGetByKeywordResponse(
 	String nextCursorId,
 	int totalCount,
-	int totalCommentCount,
+	long totalCommentCount,
 	List<VoteSummary> votes
 ) {
 	public static VoteGetByKeywordResponse from(final VoteGetByKeywordServiceResponse response) {
