+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/vote/implementation/VoteReader.java
@@ -96,7 +96,7 @@ public class VoteReader {
 	}
 
 	@Transactional(readOnly = true)
-	public int countByKeyword(final String keyword) {
+	public long countByKeyword(final String keyword) {
 		return voteRepository.countByKeyword(keyword);
 	}
