+++ bucketback-domain/src/main/java/com/programmers/bucketback/domains/vote/repository/VoteRepositoryForCursor.java
@@ -18,5 +18,5 @@ public interface VoteRepositoryForCursor {
 		final int pageSize
 	);
 
-	int countByKeyword(final String keyword);
+	Long countByKeyword(final String keyword);
 }
